+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Unreleased
+
+### Added
+- `SessionKey.Decrypt() and SessionKey.DecryptAndVerify`, now suport the decryption of AEAD encrypted 
+data packets (packet type 20).
+
 ## [2.4.7] 2022-04-27
 
 ### Changed